:truck: Move Locations component into new pattern of model based component folders.
