+++ web/src/modules/travel-authorizations/components/edit-my-travel-authorization-details-page/PurposeFormCard.vue
@@ -113,7 +113,7 @@
 <script>
 import { mapState, mapActions, mapGetters } from "vuex"
 
-import LocationsAutocomplete from "@/components/LocationsAutocomplete"
+import LocationsAutocomplete from "@/components/locations/LocationsAutocomplete"
 
 export default {
   name: "PurposeFormCard",
